Fix warnings

+++ src/map_parser/map_setup.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   map_setup.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kmordaun <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: clovell <clovell@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/21 17:00:20 by kmordaun          #+#    #+#             */
-/*   Updated: 2023/11/29 14:20:04 by kmordaun         ###   ########.fr       */
+/*   Updated: 2024/01/27 05:32:55 by clovell          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,7 +59,7 @@ t_map	map_init(t_map *map, char *map_str, t_world *world)
 	int		index;
 
 	raw_map_file = ft_lst_readfile(map_str);
-	if ((int)raw_map_file->content <= 1 || raw_map_file == NULL)
+	if ((int64_t)raw_map_file->content <= 1 || raw_map_file == NULL)
 		error_return("File Invalid", 1, 1, NULL);
 	index = 0;
 	remove_empty_lines(&raw_map_file);
